completed feature for editing and adding players

+++ OpenData.Basetball.AbaLeague.Application/DTOs/Player/PlayerDto.cs
@@ -12,7 +12,7 @@ namespace OpenData.Basketball.AbaLeague.Application.DTOs.Player
         DateTime DateOfBirth,
         Basetball.AbaLeague.Domain.Entities.Country Nationality);
 
-    public record PlayerDTO(int Id, string Name, PositionEnum Position, decimal Height,
+    public record PlayerDTO(int Id, string Name, PositionEnum Position, int Height,
         DateTime DateOfBirth, int CountryId);
 
     public record PlayerResponse(IEnumerable<PlayerDTO> Players);
